+++ src/Checkout/Contracts/CouponEvent.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Vanilo\Checkout\Contracts;
+
+interface CouponEvent extends CheckoutEvent
+{
+    public function getCouponCode(): string;
+}