+++ src/Checkout/Events/CouponUtilized.php
@@ -3,17 +3,17 @@
 declare(strict_types=1);
 
 /**
- * Contains the CouponAdded event class.
+ * Contains the CouponUtilized Event class
  *
  * @copyright   Copyright (c) 2024 Vanilo UG
  * @author      Attila Fulop
  * @license     MIT
- * @since       2024-07-30
+ * @since       2024-08-24
  *
  */
 
 namespace Vanilo\Checkout\Events;
 
-class CouponAdded extends BaseCouponEvent
+class CouponUtilized extends BaseCouponEvent
 {
 }