+++ src/Checkout/Changelog.md
@@ -13,6 +13,10 @@
   - `hasAnyCoupon()`
   - `getPromotionsAmount()`
   - `setPromotionsAmount()`
+- Added the following events:
+  - `CouponAdded` (to checkout)
+  - `CouponRemoved` (from checkout)
+  - `CouponUtilized` (after a successful checkout converted to an order)
 
 ## 4.1.0
 ##### 2024-07-11